Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.1.0
qrcode[pil]==7.4.2
websockets==12.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 